chore: restructor service streams mod
