+++ vpn/src/server.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use tokio::net::TcpListener;
 use tracing::{info, level_filters::LevelFilter};
 use tracing_subscriber::{fmt::Layer, layer::SubscriberExt, util::SubscriberInitExt, Layer as _};
-// use vpn::{ProstServerStream, TlsServerAcceptor, YamuxCtrl};
+use vpn::server::run_tcp_server;
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -14,12 +14,10 @@ async fn main() -> Result<()> {
     info!("Vpn server listening on {}", listener.local_addr()?);
 
     loop {
-        let (_stream, addr) = listener.accept().await?;
+        let (stream, addr) = listener.accept().await?;
         info!("Vpn server {:?} connected", addr);
         tokio::spawn(async move {
-            // let stream = ProstServerStream::new(stream.compat(), streams_clone);
-            // stream.process().await.unwrap();
-            // Ok(())
+            run_tcp_server(stream).await.expect("run tcp server error");
         });
     }
 }
